Revert changes to the account pool's rebuild functionality
In a situation where an account chain continuously has uncommitted blocks in the account pool, the "restabilization" of the memdbManager will eventually result in a massive amount of wrapped databases inside wrapped databases inside wrapped databases in the memdbManager's versions map. This is because the versions map never gets fully cleared and rebuilt from scratch when the Restabilize method is used. These wrapped databases inside wrapped databases can cause significant issues in read performance when reading from the unconfirmed account database.

Because of this the changes to the rebuild functionality have been reverted, even though this issue is only an edge case. More consideration is needed on how to improve the account pool's rebuild performance, while minimizing unwanted side effects.